test: `MP3FileTrimmer` 테스트 추가
